// frontend/app/services/availability/types.gen.ts
export type CubicleDto = {
    /**
     * Id
     */
    id: number;
    /**
     * Toilet Id
     */
    toilet_id: number;
    /**
     * Occupied
     */
    occupied?: boolean;
    /**
     * Toilet Roll Percentage
     */
    toilet_roll_percentage?: number;
};